edit post page

+++ controllers/api/comment-routes.js
@@ -1,5 +1,7 @@
 const router = require('express').Router();
 const { Comment } = require('../../models');
+const withAuth = require('../../utils/auth');
+
 
 router.get('/', (req, res) => {
   Comment.findAll()
@@ -10,8 +12,7 @@ router.get('/', (req, res) => {
   });
 });
 
-router.post('/', (req, res) => {
-  router.post('/', (req, res) => {
+  router.post('/', withAuth, (req, res) => {
     // check the session
     if (req.session) {
       Comment.create({
@@ -27,9 +28,10 @@ router.post('/', (req, res) => {
         });
     }
   });
-});
 
-router.delete('/:id', (req, res) => {
+
+router.delete('/:id', withAuth, (req, res) => {
+ if (req.session) {
   Comment.destroy({
     where: {
       id: req.params.id
@@ -46,6 +48,7 @@ router.delete('/:id', (req, res) => {
       console.log(err);
       res.status(500).json(err);
     });
+  }
 });
 
 module.exports = router;
